added javadocs + more methods to DynamoGearProvider, plus additional little fixes detailed in code review
