+++ caampr-backend/src/main/java/com/jelistan/caampr/lambda/dagger/ProviderModule.java
@@ -5,7 +5,7 @@ import com.amazonaws.auth.DefaultAWSCredentialsProviderChain;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
-import com.jelistan.caampr.lambda.provider.GearListProvider;
+import com.jelistan.caampr.lambda.provider.DynamoGearProvider;
 import com.jelistan.caampr.lambda.provider.GearProvider;
 import dagger.Module;
 import dagger.Provides;
@@ -18,7 +18,7 @@ public class ProviderModule {
     @Provides
     @Singleton
     GearProvider provideGearProvider(DynamoDBMapper mapper) {
-        return new GearListProvider(mapper);
+        return new DynamoGearProvider(mapper);
     }
 
     @Provides
